refactor(core): introduce startup_init.c
[no changelog]
