+++ core/embed/trezorhal/stm32u5/startup_init.c
