+++ core/site_scons/models/stm32f4_common.py
@@ -51,10 +51,10 @@ def stm32f4_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32f4/lowlevel.c",
         "embed/trezorhal/stm32f4/monoctr.c",
         "embed/trezorhal/stm32f4/mpu.c",
-        "embed/trezorhal/stm32f4/platform.c",
         "embed/trezorhal/stm32f4/random_delays.c",
         "embed/trezorhal/stm32f4/rng.c",
         "embed/trezorhal/stm32f4/secret.c",
+        "embed/trezorhal/stm32f4/startup_init.c",
         "embed/trezorhal/stm32f4/syscall.c",
         "embed/trezorhal/stm32f4/syscall_dispatch.c",
         "embed/trezorhal/stm32f4/syscall_probe.c",
