+++ core/site_scons/models/stm32u5_common.py
@@ -61,11 +61,11 @@ def stm32u5_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32u5/hash_processor.c",
         "embed/trezorhal/stm32u5/monoctr.c",
         "embed/trezorhal/stm32u5/mpu.c",
-        "embed/trezorhal/stm32u5/platform.c",
         "embed/trezorhal/stm32u5/random_delays.c",
         "embed/trezorhal/stm32u5/rng.c",
         "embed/trezorhal/stm32u5/secret.c",
         "embed/trezorhal/stm32u5/secure_aes.c",
+        "embed/trezorhal/stm32u5/startup_init.c",
         "embed/trezorhal/stm32u5/syscall.c",
         "embed/trezorhal/stm32u5/syscall_dispatch.c",
         "embed/trezorhal/stm32u5/syscall_probe.c",
